requirements.txt
discord.py==2.0.0
yt-dlp==2024.1.0
FFmpeg==5.1.2